# account-service/doc/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "List of APIs for Micro Service",
        "title": "Micro Service API Document",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api/v1",
    "paths": {
        "/adminLogin": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Admin Login API",
                "parameters": [
                    {
                        "description": "Login Data",
                        "name": "tenant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TokenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
...
        "/tokenRefresh": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Token Refresh API",
                "parameters": [
                    {
                        "description": "TokenRefresh Data",
                        "name": "tenant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TokenRefresh"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.TokenResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List all existing users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List all existing users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User Email",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Go to a specific page number. Start with 1",
                        "name": "pageNumber",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Go to specific record number. Start with 1. It will override the pageNumber if provided",
                        "name": "pageOffset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page size for the data",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Page order. Eg: name desc,createdAt desc",
                        "name": "pageOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error403"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error500"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add user",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update(Patch) user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User Data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/swagger.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/swagdto.Error404"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.LoginForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@mail.com"
                },
                "password": {
                    "type": "string",
                    "example": "john123"
                }
            }
        },
...
        "model.TokenRefresh": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGcIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MTUzNjA2MzQsInJlZfdXVpZCI6ImIyY2FyZWMzNnVzZXJfaWQiOjF9.hv_4EzMc1HG9mJZCGNk0UKnqTveOAjtgIO9Za4tHDBY"
                }
            }
        },
        "model.UserDto": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2021-02-02T02:52:24Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@mail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "User 1"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2021-02-02T02:52:24Z"
                }
            }
        },
        "model.UserForm": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@mail.com"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "name": {
                    "type": "string",
                    "example": "User 1"
                },
                "password": {
                    "type": "string",
                    "example": "Pass!23"
                }
            }
        },
...
        "swagdto.Error403": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/swagdto.ErrorForbidden"
                },
                "status": {
                    "type": "integer",
                    "example": 403
                }
            }
        },
        "swagdto.Error404": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/swagdto.ErrorNotFound"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                }
            }
        },
        "swagdto.Error500": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "error": {
                    "$ref": "#/definitions/swagdto.ErrorInternalError"
                },
                "status": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "swagdto.ErrorForbidden": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ACCESS_DENIED"
                },
                "message": {
                    "type": "string",
                    "example": "Forbidden"
                }
            }
        },
        "swagdto.ErrorInternalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INTERNAL_SERVER_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "Internal server error"
                }
            }
        },
        "swagdto.ErrorNotFound": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "Resource not found"
                }
            }
        },
...
        "swagger.TokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.UserTokenData"
                },
                "error": {
                    "type": "object"
                },
                "requestId": {
                    "type": "string",
                    "example": "3b6272b9-1ef1-45e0"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "swagger.UserListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.UserSampleListData"
                },
                "error": {
                    "type": "object"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "swagger.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/swagger.UserSampleData"
                },
                "error": {
                    "type": "object"
                },
                "status": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "swagger.UserSampleData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/model.UserDto"
                }
            }
        },
        "swagger.UserSampleListData": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserDto"
                    }
                }
            }
        },
        "swagger.UserTokenData": {
            "type": "object",
            "properties": {
                "token": {
                    "$ref": "#/definitions/swagger.UserTokenModel"
                }
            }
        },
        "swagger.UserTokenModel": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NXVCJ9.eyJhY2Nlc30IDIiLCJ0eXBlIjoiY2xpZW50In0.XBjAxzruIT"
                },
                "accessTokenExpiry": {
                    "type": "string",
                    "example": "2021-04-24T08:40:59Z"
                },
                "refreshToken": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI54XVCJ54.eyJhY2Nlc30IDIiLCJ0eXBlIjoiY2xpZW5045.XBjAxzru45"
                },
                "refreshTokenExpiry": {
                    "type": "string",
                    "example": "2021-04-21T12:40:59Z"
                },
                "tokenType": {
                    "type": "string",
                    "example": "bearer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}